+++ src/Core/Dfe.Complete.Domain/Entities/ConversionTasksData.cs
@@ -206,6 +206,14 @@ public class ConversionTasksData : BaseAggregateRoot, IEntity<TaskDataId>
 
     public bool? TenancyAtWillSaveSigned { get; set; }
 
+    public bool? TenancyAtWillBeingUsed { get; set; }
+
+    public bool? TenancyAtWillLicenceToOccupyBeingUsed { get; set; }
+
+    public bool? TenancyAtWillReceived { get; set; }
+
+    public bool? TenancyAtWillCleared { get; set; }
+
     public bool? ShareInformationEmail { get; set; }
 
     public bool? RedactAndSendSendSolicitors { get; set; }